Fix Readme to match this fork's goals, etc. from several PRs.
Got PR's pulled in from several forks off of MY fork here and the code
there is GOOD (i.e. if you took Delphi 10.x and built it, it'd work
quite correctly and handles a lot of the good work on Ashes of Avalon)
and needed to update the result to match my intended goals for this
as either the ultimate central repo for the efforts or to feed back
into SteveNew's as that function.